[bug] - Fixing price type
price shouldn't be a float.
Switching it to int

+++ scripts/buyhousing/main.py
@@ -68,7 +68,7 @@ class Propertyinfo():
     city        : str = None
     state       : str = None
     url         : str = None
-    price       : float = None
+    price       : int = None
     source      : str = None
     status      : str = None
     img_url     : str = None
